Issue #77 - Create a helpers module
